apps/backend/core-api/requirements.txt
python-dotenv
python-dateutil
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Limiter
Flask-WTF
Flask-JWT-Extended
fastapi 
uvicorn 
pydantic
pydantic-settings
sqlmodel
authlib
unidecode
gunicorn
psycopg2
google-cloud-storage
google-cloud-scheduler
google-cloud-pubsub
google-cloud-tasks
cryptography
pywebpush
rich